backend/threat_intel/requirements.txt
Django>=6.0.3
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery
redis
django-cors-headers
channels
channels-redis
gunicorn
uvicorn
yara-python
requests
python-dotenv
joblib
scikit-learn
numpy
pandas
gitpython
